Add onboarding pages and flow (#6)
This PR adds onboarding pages (avatar creation and results with a
loading state). Additionally, mock data has been added to the API folder
to improve visual representation for now and optimize the code for
easier integration with real data in the future.
- Developed onboarding pages and UI components.
- Implemented transitions between avatar creation, loading and results
pages.
- Added mock API data for testing and visualization purposes.
- Added user authentication and onboarding flow states to routing, with
a proper fallback functionality for each state (with mock data for now).